blog/requirements.txt
feedparser
requests
python-dotenv
langgraph
langchain-core
langchain-openai
langchain-groq
google-search-results